vktrace trim: Address review comments.
1) Fix missing NULL checks. In some cases NULL was not a possibility and
now the code reflects that.

2) Instead of using 'new' to allocate structs (which could throw an
exception) instead use malloc which should return NULL, and add in NULL
checks to ensure these are handled.

3) int -> uint64_t

4) Properly calculate memory size when VK_WHOLE_SIZE is used.

+++ vktrace/vktrace_common/vktrace_trace_packet_utils.c
@@ -61,7 +61,7 @@ uint64_t vktrace_get_unique_packet_index()
 
     vktrace_enter_critical_section(&s_packet_index_lock);
     // Store the value then increment within the critical section to avoid race conditions.
-    int index = s_packet_index;
+    uint64_t index = s_packet_index;
     s_packet_index += 1;
     vktrace_leave_critical_section(&s_packet_index_lock);
